refactor: quiz trying 디렉토리 제거
