+++ app/quizzes/[id]/layout.tsx
