+++ app/quizzes/[id]/loading.tsx
